+++ smooch-action-disconnect/index.js
@@ -178,6 +178,14 @@ exports.handler = async (event) => {
       log.error('An error occurred sending message', logContext, error);
       throw error;
     }
+
+    // Flow Action Response
+    await sendFlowActionResponse({
+      logContext,
+      actionId: id,
+      subId,
+      auth: cxAuth,
+    });
   }
 
   // Flow Action Response